requirements.txt
google-cloud-pubsub>=2.18.4
fastapi
uvicorn
pydantic
